----------ASSIGNMENT 4 SUBMISSION-----------------
Thanks again for a great semester. I opted out of the optional animation this time. In all honesty i am wrecked from this last push with all the assignments and i do not have the brain power left. I am going to try and make animation into a proper function over the winter break after some rest.

However, if i have not lost all brain function, i should had everything needed working. I have tested the code quite a number of time and i always clone my submission repo and try it. I do this for any assignment i hand in for any class, as you said its just good, professional, practices. As long as this clone you wont have to deal with any more of my mass of git commits lol.

Look forward to truly meeting you and the rest of the class next semester.

+++ Game2/Game/Source/Player.cpp
@@ -29,8 +29,8 @@ void Player::Update(float deltaTime)
     MoveTheBox();
     MoveTheFucker(deltaTime);
 
-    float x = round(m_Position.x / m_pTileMap->GetTileSize().x);
-    float y = round(m_Position.y / m_pTileMap->GetTileSize().x);
+    float x = float(round(m_Position.x / m_pTileMap->GetTileSize().x));
+    float y = float(round(m_Position.y / m_pTileMap->GetTileSize().x));
 
     ImGui::SliderFloat( "X", &x, 0, 10 );
     ImGui::SliderFloat( "Y", &y, 0, 10 );
@@ -164,12 +164,12 @@ void Player::MoveTheBox()
 {
     if (m_pPlayerController->IsPushHeld() && !KeyPressed)
     {
-        PushBlocks(round(m_Position.x / m_pTileMap->GetTileSize().x), round(m_Position.y / m_pTileMap->GetTileSize().x));
+        PushBlocks(int(round(m_Position.x / m_pTileMap->GetTileSize().x)), int(round(m_Position.y / m_pTileMap->GetTileSize().x)));
         KeyPressed = true;
     }
     else if (m_pPlayerController->IsPullHeld() && !KeyPressed)
     {
-        PullBlocks(round(m_Position.x / m_pTileMap->GetTileSize().x), round(m_Position.y / m_pTileMap->GetTileSize().x));
+        PullBlocks(int(round(m_Position.x / m_pTileMap->GetTileSize().x)), int(round(m_Position.y / m_pTileMap->GetTileSize().x)));
         KeyPressed = true;
     }
     else if (!m_pPlayerController->IsPushHeld() && !m_pPlayerController->IsPullHeld())
